# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing event locations
  title: Event Management Service
  version: 1.0.0
servers:
- url: http://localhost:8081
security:
- bearerAuth: []
...
paths:
...
  /users:
    post:
      description: Доступно без авторизации. По умолчанию создается пользователь с
        ролью USER. Пользователя с ролью ADMIN создать должно быть невозможно. Должна
        быть проверена уникальность логина. Логины должны быть различные у разных
        пользователей.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Пользователь успешно зарегистрирован
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      security: []
      summary: Регистрация нового пользователя
      tags:
      - Users
...
  /users/auth:
    post:
      description: Доступно без авторизации. Получение JWT токена по данным пользователя
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
          description: "Авторизация успешна, возвращен JWT"
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      security: []
      summary: Авторизация пользователя и получение JWT
      tags:
      - Users
...
components:
...
  schemas:
    ErrorMessageResponse:
      example:
        message: Сообщение об ошибке
        detailedMessage: Более детальное сообщение об ошибке
        dateTime: 2024-12-03T10:15:30
      properties:
        message:
          description: Человеко-читаемое сообщение об ошибке
          type: string
        detailedMessage:
          description: "Более детальное сообщение об ошибке. Например, содержимое\
            \ exception"
          type: string
        dateTime:
          description: Дата и время возникновения ошибки. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
      required:
      - dateTime
      - detailedMessage
      - message
      type: object
...
    User:
      description: Данные пользователя
      example:
        id: 10
        login: user_login
        age: 21
        role: USER
      properties:
        id:
          description: Уникальный идентификатор пользователя
          format: int64
          type: integer
        login:
          description: Уникальный логин пользователя
          type: string
        age:
          description: возраст пользователя
          minimum: 0
          type: int
        role:
          description: Роль пользователя
          enum:
          - ADMIN
          - USER
          type: string
      required:
      - age
      - id
      - login
      - role
      type: object
    UserRegistration:
      description: Данные для регистрации пользователя
      example:
        login: new_user
        password: strongpassword
        age: 21
      properties:
        login:
          description: Логин пользователя. Должен быть уникальным
          type: string
        password:
          description: Пароль пользователя
          type: string
        age:
          description: Возраст пользователя
          minimum: 18
          type: int
      required:
      - age
      - login
      - password
      type: object
    JwtResponse:
      example:
        jwtToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      properties:
        jwtToken:
          description: JWT token for authenticated user
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          type: string
      type: object
    UserCredentials:
      description: Данные для авторизации пользователя. Для получения JWT-токена
      example:
        login: user123
        password: password
      properties:
        login:
          description: Логин пользователя для авторизации
          type: string
        password:
          description: Пароль пользователя
          type: string
      required:
      - login
      - password
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http